Add layer functionality and improve event handling
Implemented new event handling methods in Application.cpp,
integrated GLFW for window management, and enhanced event
formatting capabilities.
Improved event handling structure in KeyEvent and MouseEvent.
Adjusted WindowsWindow.cpp for better event callbacks and modified premake5.lua for static runtime linking and Debug build configurations.

+++ game_engine/src/Engine/Events/MouseEvent.h
@@ -46,14 +46,14 @@ namespace GameEngine {
 
 	class ENGINE_API MouseButtonEvent : public Event {
 	public:
-		inline float GetMouseButton() const { return m_Button; }
+		inline int GetMouseButton() const { return m_Button; }
 
 		EVENT_CLASS_CATEGORY(EventCategoryMouse | EventCategoryInput)
 	protected:
 		MouseButtonEvent(int button)
 			: m_Button(button) {}
 
-		float m_Button;
+		int m_Button;
 	};
 
 	class ENGINE_API MouseButtonPressedEvent : public MouseButtonEvent {
